+++ packages/core/src/helpers/calls.ts
@@ -1,6 +1,6 @@
 import { utils } from 'ethers'
 import { Call } from '../hooks/useCall'
-import { Falsy } from '../model/types'
+import { Awaited, ContractMethodNames, Falsy, TypedContract } from '../model/types'
 import { RawCall, RawCallResult } from '../providers'
 import { addressEqual } from './address'
 
@@ -12,7 +12,7 @@ export function warnOnInvalidCall(call: Call | Falsy) {
   console.warn(`Invalid contract call: address=${contract.address} method=${method} args=${args}`)
 }
 
-export function encodeCallData(call: Call | Falsy): RawCall | Falsy {
+export function encodeCallData(call: Call | Falsy, chainId: number): RawCall | Falsy {
   if (!call) {
     return undefined
   }
@@ -22,7 +22,7 @@ export function encodeCallData(call: Call | Falsy): RawCall | Falsy {
     return undefined
   }
   try {
-    return { address: contract.address, data: contract.interface.encodeFunctionData(method, args) }
+    return { address: contract.address, data: contract.interface.encodeFunctionData(method, args), chainId }
   } catch {
     warnOnInvalidCall(call)
     return undefined
@@ -32,34 +32,52 @@ export function encodeCallData(call: Call | Falsy): RawCall | Falsy {
 export function getUniqueCalls(requests: RawCall[]) {
   const unique: RawCall[] = []
   for (const request of requests) {
-    if (!unique.find((x) => addressEqual(x.address, request.address) && x.data === request.data)) {
+    if (
+      !unique.find(
+        (x) => addressEqual(x.address, request.address) && x.data === request.data && x.chainId === request.chainId
+      )
+    ) {
       unique.push(request)
     }
   }
   return unique
 }
 
-export class CallError {
-  constructor(readonly message: string) {}
-}
+/**
+ * @public
+ */
+export type CallResult<T extends TypedContract, MN extends ContractMethodNames<T>> =
+  | { value: Awaited<ReturnType<T['functions'][MN]>>; error: undefined }
+  | { value: undefined; error: Error }
+  | undefined
 
-export type CallResult = { value: any[]; error: undefined } | { value: undefined; error: CallError } | undefined
-
-export function decodeCallResult(call: Call | Falsy, result: RawCallResult): CallResult {
+export function decodeCallResult<T extends TypedContract, MN extends ContractMethodNames<T>>(
+  call: Call | Falsy,
+  result: RawCallResult
+): CallResult<T, MN> {
   if (!result || !call) {
     return undefined
   }
   const { value, success } = result
-  if (success) {
-    return {
-      value: call.contract.interface.decodeFunctionResult(call.method, value) as any[],
-      error: undefined,
+  try {
+    if (success) {
+      return {
+        value: call.contract.interface.decodeFunctionResult(call.method, value) as Awaited<
+          ReturnType<T['functions'][MN]>
+        >,
+        error: undefined,
+      }
+    } else {
+      const errorMessage: string = new utils.Interface(['function Error(string)']).decodeFunctionData('Error', value)[0]
+      return {
+        value: undefined,
+        error: new Error(errorMessage),
+      }
     }
-  } else {
-    const errorMessage: string = new utils.Interface(['function Error(string)']).decodeFunctionData('Error', value)[0]
+  } catch (error) {
     return {
       value: undefined,
-      error: new CallError(errorMessage),
+      error: error as Error,
     }
   }
 }