+++ packages/core/src/helpers/transaction.ts
@@ -1,6 +1,9 @@
 import { Falsy } from '../model/types'
 import { shortenString } from './common'
 
+/**
+ * @public
+ */
 export function shortenTransactionHash(transactionHash: string): string {
   if (transactionHash.length < 10) {
     throw new TypeError('Invalid input, transaction hash need to have at least 10 characters')
@@ -9,6 +12,9 @@ export function shortenTransactionHash(transactionHash: string): string {
   return shortenString(transactionHash)
 }
 
+/**
+ * @public
+ */
 export function shortenIfTransactionHash(transactionHash: string | Falsy): string {
   if (typeof transactionHash === 'string' && transactionHash.length > 0) {
     return shortenTransactionHash(transactionHash)