+++ packages/core/src/hooks/useTransactions.ts
@@ -1,10 +1,17 @@
 import { useMemo } from 'react'
+import { QueryParams } from '../constants/type/QueryParams'
 import { useTransactionsContext } from '../providers'
 import { useEthers } from './useEthers'
 
-export function useTransactions() {
-  const { chainId, account } = useEthers()
+/**
+ * @public
+ */
+export function useTransactions(queryParams: QueryParams = {}) {
+  const { chainId: defaultChainId, account } = useEthers()
   const { addTransaction, transactions } = useTransactionsContext()
+  const { chainId: _chainId } = queryParams
+
+  const chainId = useMemo(() => _chainId ?? defaultChainId, [_chainId, defaultChainId])
 
   const filtered = useMemo(() => {
     if (chainId === undefined || !account) {