+++ packages/core/src/providers/blockNumber/context.ts
@@ -1,7 +1,0 @@
-import { createContext, useContext } from 'react'
-
-export const BlockNumberContext = createContext<number | undefined>(undefined)
-
-export function useBlockNumber() {
-  return useContext(BlockNumberContext)
-}