+++ packages/core/src/providers/network/readonlyNetworks/context.ts
@@ -1,0 +1,8 @@
+import { createContext, useContext } from 'react'
+import { Providers } from './model'
+
+export const ReadonlyNetworksContext = createContext<Providers>({})
+
+export function useReadonlyNetworks() {
+  return useContext(ReadonlyNetworksContext)
+}