+++ packages/core/src/providers/transactions/model.ts
@@ -1,6 +1,8 @@
 import { TransactionReceipt, TransactionResponse } from '@ethersproject/providers'
-import { ChainId } from '../../constants'
 
+/**
+ * @public
+ */
 export interface StoredTransaction {
   transaction: TransactionResponse
   submittedAt: number
@@ -10,6 +12,9 @@ export interface StoredTransaction {
   originalTransaction?: TransactionResponse
 }
 
+/**
+ * @public
+ */
 export function getStoredTransactionState(transaction: StoredTransaction) {
   if (transaction.receipt) {
     return transaction.receipt.status === 0 ? 'Fail' : 'Success'
@@ -17,8 +22,11 @@ export function getStoredTransactionState(transaction: StoredTransaction) {
   return 'Mining'
 }
 
+/**
+ * @public
+ */
 export type StoredTransactions = {
-  [T in ChainId]?: StoredTransaction[]
+  [chainID: number]: StoredTransaction[]
 }
 
 export const DEFAULT_STORED_TRANSACTIONS: StoredTransactions = {}