+++ packages/core/src/testing/index.ts
@@ -1,0 +1,5 @@
+import 'mock-local-storage'
+
+export * from './utils'
+export * from './renderWeb3Hook'
+export * from './mocks'