+++ packages/docs/src/examples/EthPrice.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import { useCoingeckoPrice } from '@usedapp/coingecko'
+
+ReactDOM.render(
+    <App />,
+    document.getElementById('root')
+)
+
+export function App()  {
+    const etherPrice = useCoingeckoPrice('ethereum', 'usd')
+
+    return (<p>Ether price: {etherPrice} $</p>)
+}