+++ packages/example/src/components/Transactions/Forms.tsx
@@ -5,11 +5,12 @@ import { useContractFunction, useEtherBalance, useEthers, useTokenBalance } from
 
 import { TransactionForm } from './TransactionForm'
 
+import { Weth10 } from '../../../gen/types/Weth10'
 import WethAbi from '../../abi/Weth10.json'
 
 const wethInterface = new utils.Interface(WethAbi)
 const wethContractAddress = '0xA243FEB70BaCF6cD77431269e68135cf470051b4'
-const contract = new Contract(wethContractAddress, wethInterface)
+const contract = new Contract(wethContractAddress, wethInterface) as Weth10
 
 export const DepositEth = () => {
   const { account } = useEthers()